Flattening Component Hierarchy
