+++ FE/angular/src/app/modules/movies-search/movies/movies.component.html
@@ -21,7 +21,7 @@
         {{ movie.Plot }}
       </div>
       <div class="nav flex-footer">
-        <app-go-details [imdbId]="movie.imdbID"></app-go-details>
+        <button (click)="navigateToMovieDetail(movie.imdbID)" class="detail">View Details</button>
       </div>
     </div>
   </li>
