--- FE/angular/src/app/modules/shared/navigation/go-back/go-back.component.html
@@ -1,1 +1,0 @@
-<app-sidebar [links]="links" (linkClicked)="navigateTo()"></app-sidebar>
