--- FE/angular/src/app/modules/shared/navigation/go-details/go-details.component.html
@@ -1,1 +1,0 @@
-<button (click)="navigateTo(imdbId)" class="detail">View Details</button>
