--- FE/angular/src/app/modules/shared/navigation/go-imdb/go-imdb.component.html
@@ -1,1 +1,0 @@
-<button (click)="navigateTo(imdbId)" class="imdb">View on IMDB</button>
